Adjust functional test cases with nbdkit
With the introduction of nbdkit, certain error messages do not match anymore,
and needed to be adjusted to the new behavior.

The scratch space is not needed anymore when the source is http for the majority of the cases.
Certain test cases checked the scratch space, and those are not valid anymore.

Signed-off-by: Alice Frosi <afrosi@redhat.com>